maps1/requirements.txt
python==3.9.2
ansicon==1.89.0
asgiref==3.4.1
blessed==1.19.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.8
django-appconf==1.0.5
django-classy-tags==2.0.0
django-filter==21.1
django-imagekit==4.1.0
django-location-field==2.1.0
django-mapbox-location-field==1.7.3
django-multiselectfield==0.1.12
django-taggit==2.0.0
django-taggit-templatetags2==1.6.1
future==0.18.2
geocoder==1.38.1
gpustat==1.0.0b1
gunicorn==20.1.0
idna==3.3
jinxed==1.1.0
libmagic==1.0
nvidia-ml-py3==7.352.0
pilkit==2.0
Pillow==8.4.0
portalocker==2.3.2
psutil==5.8.0
psycopg2-binary==2.9.2
python-magic-bin==0.4.14
pytz==2021.3
pywin32==302
ratelim==0.1.6
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
wcwidth==0.2.5
whitenoise==5.3.0